+++ jgclark.Summaries/src/progress_BACKUP_97911.js
@@ -12,7 +12,11 @@ import {
   getPeriodStartEndDates,
 } from './summaryHelpers'
 import { unhyphenatedDate } from '../../helpers/dateTime'
+<<<<<<< HEAD
 import { getTagParamsFromString, rangeToString } from '../../helpers/general'
+=======
+import { rangeToString } from '../../helpers/general'
+>>>>>>> 59785e290b119a6f95922b87d0d9ea82709505b2
 
 //-------------------------------------------------------------------------------
 
@@ -44,10 +48,15 @@ function getSelectedParaIndex(): number {
  * Default to looking at week to date ("wtd") but allow month to date ("mtd") as well.
  * @author @jgclark
  *
- * @param {string?} params - can pass parameter string e.g. "{interval: 'mtd', heading: 'Progress'}"
- * @return {string?} - either return string to Template, or void to plugin
+ * @param {String} interval - currently "wtd" (week-to-date) or "mtd" (month-to-date)
  */
-export async function insertProgressUpdate(params?: string): Promise<string | void > {
+<<<<<<< HEAD
+export async function insertProgressUpdate(params?: string): Promise<string> {
+=======
+export async function insertProgressUpdate(intervalType?: string): Promise<string | void> {
+  // If no intervalType passed, default to "wtd"
+  const interval = intervalType ?? 'wtd'
+>>>>>>> 59785e290b119a6f95922b87d0d9ea82709505b2
   // Get config setting
   let config = await getConfigSettings()
   // If no params are passed, then we've been called by a plugin command (and so use defaults from config).
@@ -146,10 +155,16 @@ export async function insertProgressUpdate(params?: string): Promise<string | vo
     }
   }
 
+<<<<<<< HEAD
   if (params) {
     // this was a template command call
     return `### ${heading}: Day ${dateWithinInterval} for ${periodString}\n`
       .concat(mOutputArray.join('\n'), hOutputArray.length ? '\n' : '', hOutputArray.join('\n'))
+=======
+  if (intervalType) {
+    // this was a template call
+    return mOutputArray.join('\n').concat(hOutputArray.length ? '\n' : '', hOutputArray.join('\n'))
+>>>>>>> 59785e290b119a6f95922b87d0d9ea82709505b2
   } else {
     // this is a plugin called from inside an editor
     if (Editor == null) {
@@ -163,7 +178,11 @@ export async function insertProgressUpdate(params?: string): Promise<string | vo
       }
       // console.log(`\tinserting results to current note (${Editor.filename ?? ''}) at line ${currentLineIndex}`)
       Editor.insertHeading(
+<<<<<<< HEAD
         `${heading}: Day ${dateWithinInterval} for ${periodString}`,
+=======
+        `${config.progressHeading}: Day ${dateWithinInterval} for ${periodString}`,
+>>>>>>> 59785e290b119a6f95922b87d0d9ea82709505b2
         currentLineIndex,
         3,
       )