+++ jgclark.Summaries/src/progress_LOCAL_97911.js
@@ -44,10 +44,9 @@ function getSelectedParaIndex(): number {
  * Default to looking at week to date ("wtd") but allow month to date ("mtd") as well.
  * @author @jgclark
  *
- * @param {string?} params - can pass parameter string e.g. "{interval: 'mtd', heading: 'Progress'}"
- * @return {string?} - either return string to Template, or void to plugin
+ * @param {String} interval - currently "wtd" (week-to-date) or "mtd" (month-to-date)
  */
-export async function insertProgressUpdate(params?: string): Promise<string | void > {
+export async function insertProgressUpdate(params?: string): Promise<string> {
   // Get config setting
   let config = await getConfigSettings()
   // If no params are passed, then we've been called by a plugin command (and so use defaults from config).